Conflicts:
	management/pom.xml
	management/server/core/channel-manager/pom.xml
	management/server/core/environment-manager/environment-manager-impl/src/main/java/io/subutai/core/environment/impl/EnvironmentManagerImpl.java
	management/server/core/identity-manager/identity-manager-impl/src/main/java/io/subutai/core/identity/impl/model/RolePermissionEntity.java
	management/server/subutai-common/src/main/java/io/subutai/common/environment/Blueprint.java
	management/server/subutai-features/pom.xml

+++ management/server/core/environment-manager/environment-manager-impl/src/main/java/io/subutai/core/environment/impl/EnvironmentManagerImpl.java
@@ -1,8 +1,8 @@
 package io.subutai.core.environment.impl;
 
 
+import java.security.AccessControlException;
 import java.sql.SQLException;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Map;
@@ -42,6 +42,9 @@ import io.subutai.common.peer.ContainerHost;
 import io.subutai.common.peer.EnvironmentContainerHost;
 import io.subutai.common.peer.Peer;
 import io.subutai.common.peer.PeerException;
+import io.subutai.common.security.objects.PermissionObject;
+import io.subutai.common.security.objects.PermissionOperation;
+import io.subutai.common.security.objects.PermissionScope;
 import io.subutai.common.settings.Common;
 import io.subutai.common.tracker.TrackerOperation;
 import io.subutai.common.util.ExceptionUtil;
@@ -77,6 +80,7 @@ public class EnvironmentManagerImpl implements EnvironmentManager
     private static final Logger LOG = LoggerFactory.getLogger( EnvironmentManagerImpl.class );
 
     private static final String TRACKER_SOURCE = "Environment Manager";
+    private static final String DEFAULT_GATEWAY_TEMPLATE = "192.168.%s.1/24";
 
     private final IdentityManager identityManager;
     private final PeerManager peerManager;
@@ -167,10 +171,9 @@ public class EnvironmentManagerImpl implements EnvironmentManager
     }
 
 
-    protected Topology buildTopology( final Blueprint blueprint )
+    protected Topology buildTopology( final String environmentId, final String cdir, final Blueprint blueprint )
     {
-        Topology topology = new Topology( blueprint.getName(), blueprint.getEnvironmentId(), blueprint.getCidr(),
-                blueprint.getSshKey() );
+        Topology topology = new Topology( blueprint.getName(), environmentId, cdir, blueprint.getSshKey() );
 
 
         for ( Map.Entry<String, Set<NodeGroup>> placementEntry : blueprint.getNodeGroupsMap().entrySet() )
@@ -193,16 +196,16 @@ public class EnvironmentManagerImpl implements EnvironmentManager
     {
         Preconditions.checkNotNull( blueprint, "Invalid blueprint" );
         Preconditions.checkArgument( !Strings.isNullOrEmpty( blueprint.getName() ), "Invalid name" );
-        Preconditions.checkArgument( !Strings.isNullOrEmpty( blueprint.getCidr() ), "Invalid subnet CIDR" );
+        //        Preconditions.checkArgument( !Strings.isNullOrEmpty( blueprint.getCidr() ), "Invalid subnet CIDR" );
         Preconditions.checkArgument( !blueprint.getNodeGroups().isEmpty(), "Placement is empty" );
 
-        validateBlueprint( blueprint );
+        String cdir = calculateCdir( blueprint );
 
-        Topology topology = buildTopology( blueprint );
+        String environmentId = UUID.randomUUID().toString();
+        Topology topology = buildTopology( environmentId, cdir, blueprint );
 
         //create empty environment
-        final EnvironmentImpl environment =
-                createEmptyEnvironment( blueprint.getName(), blueprint.getCidr(), blueprint.getSshKey() );
+        final EnvironmentImpl environment = createEmptyEnvironment( blueprint.getName(), cdir, blueprint.getSshKey() );
 
         //create operation tracker
         TrackerOperation operationTracker = tracker.createTrackerOperation( TRACKER_SOURCE,
@@ -252,31 +255,42 @@ public class EnvironmentManagerImpl implements EnvironmentManager
 
     private void validateBlueprint( final Blueprint blueprint ) throws EnvironmentCreationException
     {
-        checkSubnetValidity( blueprint );
+        calculateCdir( blueprint );
     }
 
 
-    private void checkSubnetValidity( final Blueprint blueprint ) throws EnvironmentCreationException
+    private String calculateCdir( final Blueprint blueprint ) throws EnvironmentCreationException
     {
         Preconditions.checkNotNull( blueprint );
-        Preconditions.checkArgument( !Strings.isNullOrEmpty( blueprint.getCidr() ) );
 
-        //check availability of subnet
-
-        SubnetUtils subnetUtils = new SubnetUtils( blueprint.getCidr() );
-        String environmentGatewayIp = subnetUtils.getInfo().getLowAddress();
         try
         {
+            Set<String> usedGateways = new HashSet<>();
             for ( String peerId : blueprint.getNodeGroupsMap().keySet() )
             {
                 Peer peer = peerManager.getPeer( peerId );
-                if ( getUsedGateways( peer ).contains( environmentGatewayIp ) )
+                usedGateways.addAll( getUsedGateways( peer ) );
+            }
+
+            String environmentGatewayIp = null;
+
+            for ( int i = 1; i < 255 && environmentGatewayIp == null; i++ )
+            {
+                SubnetUtils.SubnetInfo info = new SubnetUtils( String.format( DEFAULT_GATEWAY_TEMPLATE, i ) ).getInfo();
+
+                String gw = info.getLowAddress();
+
+                if ( !usedGateways.contains( gw ) )
                 {
-                    throw new EnvironmentCreationException(
-                            String.format( "Subnet %s is already used on peerId %s", environmentGatewayIp,
-                                    peer.getName() ) );
+                    environmentGatewayIp = info.getCidrSignature();
                 }
             }
+
+            if ( environmentGatewayIp == null )
+            {
+                throw new EnvironmentCreationException( "Could not determine subnet cdir." );
+            }
+            return environmentGatewayIp;
         }
         catch ( PeerException e )
         {
@@ -301,33 +315,36 @@ public class EnvironmentManagerImpl implements EnvironmentManager
 
     @RolesAllowed( "Environment-Management|A|Write" )
     @Override
-    public Set<EnvironmentContainerHost> growEnvironment( final Blueprint blueprint, final boolean async )
+    public Set<EnvironmentContainerHost> growEnvironment( final String environmentId, final Blueprint blueprint,
+                                                          final boolean async )
             throws EnvironmentModificationException, EnvironmentNotFoundException
     {
         Preconditions.checkNotNull( blueprint, "Invalid blueprint" );
-        Preconditions.checkArgument( !Strings.isNullOrEmpty( blueprint.getEnvironmentId() ), "Invalid environment id" );
+        Preconditions.checkArgument( !Strings.isNullOrEmpty( environmentId ), "Invalid environment id" );
         Preconditions.checkArgument( !blueprint.getNodeGroups().isEmpty(), "Placement is empty" );
         TrackerOperation operationTracker = tracker.createTrackerOperation( TRACKER_SOURCE,
-                String.format( "Growing environment %s", blueprint.getEnvironmentId() ) );
-        return growEnvironment( blueprint, async, true, operationTracker );
+                String.format( "Growing environment %s", environmentId ) );
+        return growEnvironment( environmentId, blueprint, async, true, operationTracker );
     }
 
 
     @RolesAllowed( "Environment-Management|A|Write" )
-    private Set<EnvironmentContainerHost> growEnvironment( final Blueprint blueprint, final boolean async,
-                                                           final boolean checkAccess,
+    private Set<EnvironmentContainerHost> growEnvironment( final String environmentId, final Blueprint blueprint,
+                                                           final boolean async, final boolean checkAccess,
                                                            TrackerOperation operationTracker )
             throws EnvironmentModificationException, EnvironmentNotFoundException
     {
 
         Preconditions.checkNotNull( blueprint, "Invalid blueprint" );
-        Preconditions.checkArgument( !Strings.isNullOrEmpty( blueprint.getEnvironmentId() ), "Invalid environment id" );
+        Preconditions.checkArgument( !Strings.isNullOrEmpty( environmentId ), "Invalid environment id" );
         Preconditions.checkArgument( !blueprint.getNodeGroups().isEmpty(), "Placement is empty" );
 
-        final Topology topology = buildTopology( blueprint );
+        final EnvironmentImpl environment = ( EnvironmentImpl ) loadEnvironment( environmentId, checkAccess );
 
-        final EnvironmentImpl environment =
-                ( EnvironmentImpl ) loadEnvironment( blueprint.getEnvironmentId(), checkAccess );
+        String cdir = environment.getSubnetCidr();
+
+        final Topology topology = buildTopology( environmentId, cdir, blueprint );
+
 
         if ( environment.getStatus() == EnvironmentStatus.UNDER_MODIFICATION )
         {
@@ -547,6 +564,18 @@ public class EnvironmentManagerImpl implements EnvironmentManager
 
         final EnvironmentImpl environment = ( EnvironmentImpl ) loadEnvironment( environmentId, checkAccess );
 
+        User activeUser = identityManager.getActiveUser();
+
+        final boolean deleteAll = identityManager
+                .isUserPermitted( activeUser, PermissionObject.EnvironmentManagement, PermissionScope.ALL_SCOPE,
+                        PermissionOperation.Delete );
+
+        if (!( deleteAll || environment.getUserId().equals( activeUser.getId() ) ))
+        {
+            throw new AccessControlException( "You have not enough permissions." );
+        }
+
+
         if ( environment.getStatus() == EnvironmentStatus.UNDER_MODIFICATION )
         {
             operationTracker.addLogFailed( String.format( "Environment status is %s", environment.getStatus() ) );
@@ -628,9 +657,21 @@ public class EnvironmentManagerImpl implements EnvironmentManager
 
         Environment environment = loadEnvironment( environmentId, checkAccess );
 
-        environmentDataService.remove( ( EnvironmentImpl ) environment );
+        User activeUser = identityManager.getActiveUser();
 
-        notifyOnEnvironmentDestroyed( environmentId );
+        final boolean deleteAll = identityManager
+                .isUserPermitted( activeUser, PermissionObject.EnvironmentManagement, PermissionScope.ALL_SCOPE,
+                        PermissionOperation.Delete );
+
+        if ( deleteAll || environment.getUserId().equals( activeUser.getId() ) )
+        {
+            environmentDataService.remove( ( EnvironmentImpl ) environment );
+            notifyOnEnvironmentDestroyed( environmentId );
+        }
+        else
+        {
+            throw new AccessControlException( "You have not enough permissions." );
+        }
     }
 
 
@@ -638,28 +679,25 @@ public class EnvironmentManagerImpl implements EnvironmentManager
     @Override
     public Set<Environment> getEnvironments()
     {
-        Set<Environment> environments = Sets.newHashSet();
-        environments.addAll( environmentDataService.getAll() );
+        User activeUser = identityManager.getActiveUser();
 
-        for ( Environment environment : environments )
+        final boolean viewAll = identityManager
+                .isUserPermitted( activeUser, PermissionObject.EnvironmentManagement, PermissionScope.ALL_SCOPE,
+                        PermissionOperation.Read );
+
+
+        Set<Environment> environments = new HashSet<>();
+        for ( Environment environment : environmentDataService.getAll() )
         {
-            setEnvironmentTransientFields( environment );
-            setContainersTransientFields( environment );
+            if ( viewAll || environment.getUserId().equals( activeUser.getId() ) )
+            {
+                environments.add( environment );
+
+                setEnvironmentTransientFields( environment );
+                setContainersTransientFields( environment );
+            }
         }
 
-        //@todo remove temporarily frontend-ui
-//        if ( !isUserAdmin() )
-//        {
-//            Long userId = getUserId();
-//            for ( Iterator<Environment> iterator = environments.iterator(); iterator.hasNext(); )
-//            {
-//                final Environment environment = iterator.next();
-//                if ( !Objects.equals( environment.getUserId(), userId ) )
-//                {
-//                    iterator.remove();
-//                }
-//            }
-//        }
 
         return environments;
     }
@@ -1140,6 +1178,7 @@ public class EnvironmentManagerImpl implements EnvironmentManager
         EnvironmentImpl environment =
                 new EnvironmentImpl( name, subnetCidr, sshKey, getUserId(), peerManager.getLocalPeerInfo().getId() );
 
+        environment.setUserId( identityManager.getActiveUser().getId() );
         environment = saveOrUpdate( environment );
 
         setEnvironmentTransientFields( environment );
